feat(role): add role in schema and data initialization
In schema.sql, add the role table with a primary key on `id` and a
unique index on `name`.
Modify the `user` table to include a foreign key on the `role` field,
referencing the `id` field of the `role` table.

In data.sql, modify users insertions and insert initial roles using
INSERT INTO... SELECT... WHERE NOT EXISTS to prevent duplicate errors.